Split up direct-messages module
