+++ test/utils/direct-messages-test.ts
@@ -15,25 +15,19 @@ limitations under the License.
 */
 
 import { mocked } from "jest-mock";
-import {
-    ClientEvent,
-    EventType,
-    KNOWN_SAFE_ROOM_VERSION,
-    MatrixClient,
-    Room,
-} from "matrix-js-sdk/src/matrix";
+import { ClientEvent, MatrixClient, Room } from "matrix-js-sdk/src/matrix";
 
 import DMRoomMap from "../../src/utils/DMRoomMap";
-import { createTestClient, makeMembershipEvent } from "../test-utils";
+import { createTestClient } from "../test-utils";
 import { LocalRoom, LocalRoomState, LOCAL_ROOM_ID_PREFIX } from "../../src/models/LocalRoom";
 import * as dmModule from "../../src/utils/direct-messages";
 import dis from "../../src/dispatcher/dispatcher";
 import { Action } from "../../src/dispatcher/actions";
 import { MatrixClientPeg } from "../../src/MatrixClientPeg";
-import { privateShouldBeEncrypted } from "../../src/utils/rooms";
-import { Member } from "../../src/utils/direct-messages";
-import { canEncryptToAllUsers } from "../../src/createRoom";
 import { waitForRoomReadyAndApplyAfterCreateCallbacks } from "../../src/utils/local-room";
+import { findDMRoom } from "../../src/utils/dm/findDMRoom";
+import { createDmLocalRoom } from "../../src/utils/dm/createDmLocalRoom";
+import { startDm } from "../../src/utils/dm/startDm";
 
 jest.mock("../../src/utils/rooms", () => ({
     ...(jest.requireActual("../../src/utils/rooms") as object),
@@ -49,34 +43,25 @@ jest.mock("../../src/utils/local-room", () => ({
     waitForRoomReadyAndApplyAfterCreateCallbacks: jest.fn(),
 }));
 
-function assertLocalRoom(room: LocalRoom, targets: Member[], encrypted: boolean) {
-    expect(room.roomId).toBe(LOCAL_ROOM_ID_PREFIX + "t1");
-    expect(room.name).toBe(targets.length ? targets[0].name : "Empty Room");
-    expect(room.encrypted).toBe(encrypted);
-    expect(room.targets).toEqual(targets);
-    expect(room.getMyMembership()).toBe("join");
+jest.mock("../../src/utils/dm/findDMForUser", () => ({
+    findDMForUser: jest.fn(),
+}));
 
-    const roomCreateEvent = room.currentState.getStateEvents(EventType.RoomCreate)[0];
-    expect(roomCreateEvent).toBeDefined();
-    expect(roomCreateEvent.getContent()["room_version"]).toBe(KNOWN_SAFE_ROOM_VERSION);
+jest.mock("../../src/utils/dm/findDMRoom", () => ({
+    findDMRoom: jest.fn(),
+}));
 
-    // check that the user and all targets are joined
-    expect(room.getMember("@userId:matrix.org").membership).toBe("join");
-    targets.forEach((target: Member) => {
-        expect(room.getMember(target.userId).membership).toBe("join");
-    });
+jest.mock("../../src/utils/dm/createDmLocalRoom", () => ({
+    createDmLocalRoom: jest.fn(),
+}));
 
-    if (encrypted) {
-        const encryptionEvent = room.currentState.getStateEvents(EventType.RoomEncryption)[0];
-        expect(encryptionEvent).toBeDefined();
-    }
-}
+jest.mock("../../src/utils/dm/startDm", () => ({
+    startDm: jest.fn(),
+}));
 
 describe("direct-messages", () => {
     const userId1 = "@user1:example.com";
     const member1 = new dmModule.DirectoryMember({ user_id: userId1 });
-    const userId2 = "@user2:example.com";
-    const member2 = new dmModule.ThreepidMember("user2");
     let room1: Room;
     let localRoom: LocalRoom;
     let dmRoomMap: DMRoomMap;
@@ -108,117 +93,14 @@ describe("direct-messages", () => {
         jest.setSystemTime(new Date(2022, 7, 4, 11, 12, 30, 42));
     });
 
-    describe("findDMForUser", () => {
-        describe("for an empty DM room list", () => {
-            beforeEach(() => {
-                mocked(dmRoomMap.getDMRoomsForUserId).mockReturnValue([]);
-            });
-
-            it("should return undefined", () => {
-                expect(dmModule.findDMForUser(mockClient, userId1)).toBeUndefined();
-            });
-        });
-
-        describe("for some rooms", () => {
-            let room2: LocalRoom;
-            let room3: Room;
-            let room4: Room;
-            let room5: Room;
-
-            beforeEach(() => {
-                room1.currentState.setStateEvents([
-                    makeMembershipEvent(room1.roomId, userId1, "join"),
-                    makeMembershipEvent(room1.roomId, userId2, "join"),
-                ]);
-
-                // this should not be a DM room because it is a local room
-                room2 = new LocalRoom("!room2:example.com", mockClient, userId1);
-                room2.getLastActiveTimestamp = () => 100;
-
-                room3 = new Room("!room3:example.com", mockClient, userId1);
-                room3.getMyMembership = () => "join";
-                room3.currentState.setStateEvents([
-                    makeMembershipEvent(room3.roomId, userId1, "join"),
-                    makeMembershipEvent(room3.roomId, userId2, "join"),
-                ]);
-
-                // this should not be a DM room because it has only one joined user
-                room4 = new Room("!room4:example.com", mockClient, userId1);
-                room4.getMyMembership = () => "join";
-                room4.currentState.setStateEvents([
-                    makeMembershipEvent(room4.roomId, userId1, "invite"),
-                    makeMembershipEvent(room4.roomId, userId2, "join"),
-                ]);
-
-                // this should not be a DM room because it has no users
-                room5 = new Room("!room5:example.com", mockClient, userId1);
-                room5.getLastActiveTimestamp = () => 100;
-
-                mocked(mockClient.getRoom).mockImplementation((roomId: string) => {
-                    return {
-                        [room1.roomId]: room1,
-                        [room2.roomId]: room2,
-                        [room3.roomId]: room3,
-                        [room4.roomId]: room4,
-                        [room5.roomId]: room5,
-                    }[roomId];
-                });
-
-                mocked(dmRoomMap.getDMRoomsForUserId).mockReturnValue([
-                    room1.roomId,
-                    room2.roomId,
-                    room3.roomId,
-                    room4.roomId,
-                    room5.roomId,
-                ]);
-            });
-
-            it("should find a room ordered by last activity 1", () => {
-                room1.getLastActiveTimestamp = () => 2;
-                room3.getLastActiveTimestamp = () => 1;
-
-                expect(dmModule.findDMForUser(mockClient, userId1)).toBe(room1);
-            });
-
-            it("should find a room ordered by last activity 2", () => {
-                room1.getLastActiveTimestamp = () => 1;
-                room3.getLastActiveTimestamp = () => 2;
-
-                expect(dmModule.findDMForUser(mockClient, userId1)).toBe(room3);
-            });
-        });
-    });
-
-    describe("findDMRoom", () => {
-        it("should return the room for a single target with a room", () => {
-            jest.spyOn(dmModule, "findDMForUser").mockReturnValue(room1);
-            expect(dmModule.findDMRoom(mockClient, [member1])).toBe(room1);
-        });
-
-        it("should return null for a single target without a room", () => {
-            jest.spyOn(dmModule, "findDMForUser").mockReturnValue(null);
-            expect(dmModule.findDMRoom(mockClient, [member1])).toBeNull();
-        });
-
-        it("should return the room for 2 targets with a room", () => {
-            mocked(dmRoomMap.getDMRoomForIdentifiers).mockReturnValue(room1);
-            expect(dmModule.findDMRoom(mockClient, [member1, member2])).toBe(room1);
-        });
-
-        it("should return null for 2 targets without a room", () => {
-            mocked(dmRoomMap.getDMRoomForIdentifiers).mockReturnValue(null);
-            expect(dmModule.findDMRoom(mockClient, [member1, member2])).toBeNull();
-        });
-    });
-
     describe("startDmOnFirstMessage", () => {
         describe("if no room exists", () => {
             beforeEach(() => {
-                jest.spyOn(dmModule, "findDMRoom").mockReturnValue(null);
+                mocked(findDMRoom).mockReturnValue(null);
             });
 
             it("should create a local room and dispatch a view room event", async () => {
-                jest.spyOn(dmModule, "createDmLocalRoom").mockResolvedValue(localRoom);
+                mocked(createDmLocalRoom).mockResolvedValue(localRoom);
                 const room = await dmModule.startDmOnFirstMessage(mockClient, [member1]);
                 expect(room).toBe(localRoom);
                 expect(dis.dispatch).toHaveBeenCalledWith({
@@ -232,7 +114,7 @@ describe("direct-messages", () => {
 
         describe("if a room exists", () => {
             beforeEach(() => {
-                jest.spyOn(dmModule, "findDMRoom").mockReturnValue(room1);
+                mocked(findDMRoom).mockReturnValue(room1);
             });
 
             it("should return the room and dispatch a view room event", async () => {
@@ -249,72 +131,19 @@ describe("direct-messages", () => {
         });
     });
 
-    describe("createDmLocalRoom", () => {
-        describe("when rooms should be encrypted", () => {
-            beforeEach(() => {
-                mocked(privateShouldBeEncrypted).mockReturnValue(true);
-            });
-
-            it("should create an unencrypted room for 3PID targets", async () => {
-                const room = await dmModule.createDmLocalRoom(mockClient, [member2]);
-                expect(mockClient.store.storeRoom).toHaveBeenCalledWith(room);
-                assertLocalRoom(room, [member2], false);
-            });
-
-            describe("for MXID targets with encryption available", () => {
-                beforeEach(() => {
-                    mocked(canEncryptToAllUsers).mockResolvedValue(true);
-                });
-
-                it("should create an encrypted room", async () => {
-                    const room = await dmModule.createDmLocalRoom(mockClient, [member1]);
-                    expect(mockClient.store.storeRoom).toHaveBeenCalledWith(room);
-                    assertLocalRoom(room, [member1], true);
-                });
-            });
-
-            describe("for MXID targets with encryption unavailable", () => {
-                beforeEach(() => {
-                    mocked(canEncryptToAllUsers).mockResolvedValue(false);
-                });
-
-                it("should create an unencrypted room", async () => {
-                    const room = await dmModule.createDmLocalRoom(mockClient, [member1]);
-                    expect(mockClient.store.storeRoom).toHaveBeenCalledWith(room);
-                    assertLocalRoom(room, [member1], false);
-                });
-            });
-        });
-
-        describe("if rooms should not be encrypted", () => {
-            beforeEach(() => {
-                mocked(privateShouldBeEncrypted).mockReturnValue(false);
-            });
-
-            it("should create an unencrypted room", async () => {
-                const room = await dmModule.createDmLocalRoom(mockClient, [member1]);
-                assertLocalRoom(room, [member1], false);
-            });
-        });
-    });
-
     describe("createRoomFromLocalRoom", () => {
-        beforeEach(() => {
-            jest.spyOn(dmModule, "startDm");
-        });
-
         [LocalRoomState.CREATING, LocalRoomState.CREATED, LocalRoomState.ERROR].forEach((state: LocalRoomState) => {
             it(`should do nothing for room in state ${state}`, async () => {
                 localRoom.state = state;
                 await dmModule.createRoomFromLocalRoom(mockClient, localRoom);
                 expect(localRoom.state).toBe(state);
-                expect(dmModule.startDm).not.toHaveBeenCalled();
+                expect(startDm).not.toHaveBeenCalled();
             });
         });
 
         describe("on startDm error", () => {
             beforeEach(() => {
-                mocked(dmModule.startDm).mockRejectedValue(true);
+                mocked(startDm).mockRejectedValue(true);
             });
 
             it("should set the room state to error", async () => {
@@ -326,7 +155,7 @@ describe("direct-messages", () => {
         describe("on startDm success", () => {
             beforeEach(() => {
                 mocked(waitForRoomReadyAndApplyAfterCreateCallbacks).mockResolvedValue(room1.roomId);
-                mocked(dmModule.startDm).mockResolvedValue(room1.roomId);
+                mocked(startDm).mockResolvedValue(room1.roomId);
             });
 
             it(
